Refactor puppeteer
